:truck: Move QueueFamilyIndices class into /core/queue/
